requirements.txt
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
finvizlite==0.1.1
idna==2.10
iniconfig==1.1.1
lxml==4.6.2
multitasking==0.0.9
numpy==1.19.4
packaging==20.7
pandas==1.1.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
py-lets-be-rational==1.0.1
py-vollib==1.0.1
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
toml==0.10.2
urllib3==1.26.2
yfinance==0.1.55
